+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Docker Language Server will be documented in this file.
 
-## [Unreleased]
+## [0.10.2] - 2025-06-06
 
 ### Fixed
 
@@ -410,7 +410,8 @@ All notable changes to the Docker Language Server will be documented in this fil
   - textDocument/semanticTokens/full
     - provide syntax highlighting for Bake files
 
-[Unreleased]: https://github.com/docker/docker-language-server/compare/v0.10.1...main
+[Unreleased]: https://github.com/docker/docker-language-server/compare/v0.10.2...main
+[0.10.2]: https://github.com/docker/docker-language-server/compare/v0.10.1...v0.10.2
 [0.10.1]: https://github.com/docker/docker-language-server/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/docker/docker-language-server/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/docker/docker-language-server/compare/v0.8.0...v0.9.0